Backend/YOLO_Project/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
ultralytics
opencv-python
torch
torchvision
requests
python-multipart
paho-mqtt
